Integration tests for the image-service
Verify that we can grab a few versions (and architectures) of RHCOS from
the openshift mirror and:
1. Create full-iso, minimal-iso template
2. Stream an RHCOS image with ignition + initrd customizations embedded
   properly

Since this test could take a long time (and longer in the future as we
support more versions and architectures), we'll skip this when run using
`go test -short`. So this change also includes a new make target
`test-integration` to run this test.